Add idt::InterruptDescriptorTable (#902)
* Add lidt instruction

* Move TYPE_TSS to x86::seg

* Add idt::InterruptDescriptorTable

* Update tests

* Add x86_64 gate to set_stack_index

* Update comment

* Add x86::int::InterruptFrame

* Update changelog

+++ src/sys/pic.rs
@@ -49,8 +49,8 @@ pub fn unmask(irq: u8) {
     }
 }
 
-pub fn vector(irq: u8) -> u8 {
-    PIC_OFFSET + irq
+pub fn vector(irq: u8) -> usize {
+    (PIC_OFFSET + irq) as usize
 }
 
 pub fn eoi(irq: u8) {
